+++ lib/sqlreport/result.rb
@@ -1,17 +1,30 @@
 # frozen_string_literal: true
 
 require "csv"
+require "yaml"
 
 module Sqlreport
   # Result
   class Result
-    def initialize(query)
+    def initialize(query, db_config: false)
       @query = query
-      @connection = ActiveRecord::Base.connection
+      @db_config = db_config
       @response = nil
     end
 
-    def result
+    def connection
+      @connection ||= if @db_config
+                        ActiveRecord::Base.establish_connection(@db_config)
+                      else
+                        ActiveRecord::Base.connection
+                      end
+    end
+
+    def result(validate: true)
+      connection
+      validations = validate_input if validate
+      return validations if validations
+
       @response = @connection.exec_query(@query)
       self
     end
@@ -35,6 +48,15 @@ module Sqlreport
       data = to_csv(include_headers: include_headers, separator: separator, quote_char: quote_char)
       File.write(path, data)
       true
+    end
+
+    private
+
+    def validate_input
+      return "DELETE, UPDATE, DROP, RENAME, ALTER cannot be used in queries" \
+        if %w[DELETE UPDATE DROP RENAME ALTER].any? { |needle| @query.include? needle }
+
+      false
     end
   end
 end